tests/wpt/tests/intersection-observer/scroll-margin-non-scrolling-root.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<link rel="help" href="https://www.w3.org/TR/intersection-observer/#dom-intersectionobserver-scrollmargin">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
#scroller { width: 100px; height: 100px; overflow: hidden; background-color: gray; }
#spacer { width: 50px; height: 110px; background-color: darkgray; }
#root { height: 75; width: 75; background-color: blue; }
#target { width: 50px; height: 50px; background-color: green; }
#spacer2 { width: 10px; height: 10px; background-color: lightgray; }
</style>

<div id=scroller>
  <div id=spacer></div>
  <div id=root>
    <div id=spacer2></div>
    <div id=target></div>
  </div>
</div>

<script>
let entries = [];

window.onload = function() {
  runTestCycle(testIntersection, "Test scroll margin intersection");

  const observer = new IntersectionObserver(
    es => entries = entries.concat(es),
    {
      root: document.querySelector("#root"),
      scrollMargin: "10px"
    }
  );
  observer.observe(target);
};

function testIntersection() {
  assert_equals(entries.length, 1, "IntersectionObserverEntryCount");
  assert_true(entries[0].isIntersecting, "isIntersecting");
  assert_approx_equals(entries[0].intersectionRatio, 1.0, 0.001, "intersectionRatio");
}
</script>
